pull request fixes
